Changes Made for Image Processing Test
Edited GoProController to save each pulled image as 'image.png'. This
file will be overwritten each time an image is pulled. The file is
opened, loaded, and appended to an array outside of the Detector.detect() function.
Then the Detector.detect() function is passed items from the array as parameters.